moves vote specs to budget scoped folder
